AArch64: Refactor cross-compilation options
The existing Cross Compilation toolchains for AArch64 use environment
variables to specify optional features, such as SVE/SVE2, which are then
used to set CMake variables.

Refactor this using CMake options (boolean cache variables), so that
these can be set directly when running CMake to configure the project.
As the options are defined in the top-level CMakeLists file the
duplication between cross-compilation toolchain files is reduced.