AArch64: Separate setup of optimized primitives
Currently the intrinsics-based optimized primitives for AArch64 are
setup at the start of setupAssemblyPrimitives, rather than in the
separate, intrinsics-specific function. As this same combined setup
function is used when setting up the tests, only a specific subset of
intrinsics primitives are tested in testbench.cpp.

This patch moves the setup for AArch64 intrinsics implementations to
setupIntrinsicPrimitives, and enables separate testing for
intrinsics primitives. This change does not alter the function table
produced in x265_setup_primitives(), as the intrinsics primitives are
still setup prior to the ASM primitives.