source/CMakeLists.txt: allow setting CPU_HAS_NEON when CROSS_COMPILE_ARM
The logic in source/cmake/FindNeon.cmake is not appropriate for
cross-compilation scenarios, so in order to allow cross-compiling for
ARM, CROSS_COMPILE_ARM needs to be defined. However that currently
doesn't allow the enabling of NEON support, so let's allow the user to
manually set CPU_HAS_NEON in this case.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>